+++ src/main/java/com/ningyang/os/service/impl/SerPurchaseOrderInfoServiceImpl.java
@@ -48,7 +48,7 @@ public class SerPurchaseOrderInfoServiceImpl extends ServiceImpl<SerPurchaseOrde
             int boxCount = detailsService.boxCount(condition);
             vo.setProductNumber(boxCount);
         }
-        int total = baseMapper.selectOrderPurchaseVoPageCountByCondition(condition);
+        Long total = baseMapper.selectOrderPurchaseVoPageCountByCondition(condition);
         pageVo.setRecords(listVoTemp);
         pageVo.setTotal(total);
         pageVo.setSize(condition.getPage());
